Use larger types
size_t instead of unsigned int for size of containers
std::int64_t instead of int
std::uint64_t instead of unsigned int

+++ hdi/data/embedding.h
@@ -51,16 +51,16 @@ namespace hdi{
 
     public:
       Embedding();
-      Embedding(unsigned int num_dimensions, unsigned int num_data_points, scalar_type v = 0);
+      Embedding(size_t num_dimensions, size_t num_data_points, scalar_type v = 0);
 
       //! Clear the container
       void clear();
       //! Resize the container
-      void resize(unsigned int num_dimensions, unsigned int num_data_points, scalar_type v = 0);
+      void resize(size_t num_dimensions, size_t num_data_points, scalar_type v = 0);
       //! Return the dimensionality of the embedding
-      unsigned int numDimensions()const{return _num_dimensions;}
+      size_t numDimensions()const{return _num_dimensions;}
       //! Return the number of data points in the container
-      unsigned int numDataPoints()const{return _num_data_points;}
+      size_t numDataPoints()const{return _num_data_points;}
       //! Compute a boundinb box that contains the embedding. An offset can be provided (percentage)
       void computeEmbeddingBBox(scalar_vector_type& limits, scalar_type offset = 0, bool squared_limits = true);
 
@@ -74,20 +74,20 @@ namespace hdi{
       scalar_vector_type& getContainer(){return _embedding;}
       const scalar_vector_type& getContainer()const{return _embedding;}
 
-      inline scalar_type& dataAt(unsigned int data_point, unsigned int dimension){
+      inline scalar_type& dataAt(size_t data_point, size_t dimension){
         assert(data_point < _num_data_points);
         assert(dimension < _num_dimensions);
         return _embedding[data_point*_num_dimensions+dimension];
       }
-      inline const scalar_type& dataAt(unsigned int data_point, unsigned int dimension)const{
+      inline const scalar_type& dataAt(size_t data_point, size_t dimension)const{
         assert(data_point < _num_data_points);
         assert(dimension < _num_dimensions);
         return _embedding[data_point*_num_dimensions+dimension];
       }
 
     private:
-      unsigned int _num_dimensions;
-      unsigned int _num_data_points;
+      size_t _num_dimensions;
+      size_t _num_data_points;
       scalar_vector_type _embedding;
     };
